Only clear the collection when the debug variable is set.

+++ source/projects/main/project/importers/redirects.file.js
@@ -14,9 +14,11 @@
                 .replace(new RegExp(`[${charList}]+$`), "");
         };
 
-    try {
-        await db.collection("routes").drop();
-    } catch (err) {}
+    if (process.env.DEBUG) {
+        try {
+            await db.collection("routes").drop();
+        } catch (err) {}
+    }
 
     let newRoutesCollection = db.collection("routes");
